requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.3
black==25.1.0
blinker==1.9.0
click==8.1.8
flake8==7.2.0
Flask==3.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
python-dotenv==1.1.0
Werkzeug==3.1.3
